requirements.txt
python-dotenv
requests
slack_bolt
slack-sdk

